DT-1071 Fix timing issues with integration tests building indexes with zero offenders
With offenders we can wait until they appear in elasticsearch to know the build is complete.  Without offenders this doesn't work as we are waiting for elasticsearch to have zero offenders - it already does and this causes a race condition in the tests.  Fix by just checking that all messages have been processed i.e. there is nothing active on the index queue.